Sarthak

+++ Scripts/XML parsing.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# In[2]:
+# In[6]:
 
 
 import xmltodict
@@ -10,9 +10,43 @@ import pandas as pd
 import time
 import boto3
 import shutil
+from nltk.corpus import stopwords
+from collections import Counter
+import re
+from nltk.stem import WordNetLemmatizer
+from nltk.stem import PorterStemmer
+import nltk
+nltk.download('averaged_perceptron_tagger')
+from nltk.corpus import wordnet
 
 
-# In[3]:
+# In[7]:
+
+
+stop = stopwords.words('english')
+lemmatizer = WordNetLemmatizer()
+porter = PorterStemmer()
+
+
+# In[8]:
+
+
+# POS_TAGGER_FUNCTION : TYPE 1
+def pos_tagger(nltk_tag):
+    #print("Tag is here ",nltk_tag)
+    if nltk_tag.startswith('J'):
+        return wordnet.ADJ
+    elif nltk_tag.startswith('V'):
+        return wordnet.VERB
+    elif nltk_tag.startswith('N'):
+        return wordnet.NOUN
+    elif nltk_tag.startswith('R'):
+        return wordnet.ADV
+    else:         
+        return None
+
+
+# In[9]:
 
 
 def dataframe_generator(partition):
@@ -25,13 +59,13 @@ def dataframe_generator(partition):
     org_names=[]
     org_cities=[]
     org_countries=[]
-    
     nois = []
     abstracts=[]
-    
+    word_counts = []
     
     files = os.listdir('Unzipped/{}'.format(partition))
     for file in files: 
+        word_list = []
         with open(f'Unzipped/{partition}/{file}') as f:
             try:
                 dic=xmltodict.parse(f.read())
@@ -55,23 +89,49 @@ def dataframe_generator(partition):
 
                 nois.append(len(dic['us-patent-application']['us-bibliographic-data-application']['us-parties']['inventors']['inventor']))
                 abstracts.append(dic['us-patent-application']['abstract']['p']['#text'])
+                
+                for i in range(len(dic['us-patent-application']['claims']['claim'])):
+                    try:
+                        text = dic['us-patent-application']['claims']['claim'][i]['claim-text']['#text']
+                        
+                        all_words = re.sub(r'[^\w\s]', '', text).split()
+
+                        pos_tagged = nltk.pos_tag(all_words) 
+                        
+                        wordnet_tagged = list(map(lambda x: (x[0], pos_tagger(x[1])), pos_tagged))
+                        
+                        
+                        lemmatized_sentence = []
+                        for word, tag in wordnet_tagged:
+                            if tag is None:
+                                # if there is no available tag, append the token as is
+                                lemmatized_sentence.append(word)
+                            else:       
+                                # else use the tag to lemmatize the token
+                                lemmatized_sentence.append(lemmatizer.lemmatize(word, tag))
+                        
+                        filtered_words = [word.lower() for word in lemmatized_sentence if word.lower() not in stop]
+                        word_list.extend(filtered_words)
+                    except:
+                        continue
+                word_counts.append(Counter(word_list).items())
             except:
                 continue
-        
-    
-    df = pd.DataFrame(list(zip(doc_numbers, titles, utilities, dates, countries, org_names, org_cities, org_countries, nois, abstracts)),
-               columns =['Doc_number', 'Title', 'Type','App_Date','Country','Org_Name','Org_City','Org_Country','No_inventors','Abstract'])
+                
+    df = pd.DataFrame(list(zip(doc_numbers, titles, utilities, dates, countries, org_names, org_cities, org_countries, nois, abstracts, word_counts)),
+               columns =['Doc_number', 'Title', 'Type','App_Date','Country','Org_Name','Org_City','Org_Country','No_inventors','Abstract','Description'])
     
     df['App_Date'] = pd.to_datetime(df['App_Date'])
+    
     
     return df
 
 
-# In[4]:
+# In[10]:
 
 
-access_key_id  = '------'
-secret_access_key = '------'
+access_key_id  = ''
+secret_access_key = ''
 
 session = boto3.Session(aws_access_key_id=access_key_id, aws_secret_access_key=secret_access_key)
 resources = session.resource('s3')
@@ -80,14 +140,14 @@ s3 = session.client('s3')
 my_bucket = resources.Bucket('capstone-storage')
 
 
-# In[15]:
+# In[11]:
 
 
 response = s3.list_objects_v2(Bucket='capstone-storage')
 files = response.get("Contents")
 
 
-# In[36]:
+# In[12]:
 
 
 locs = []
@@ -97,7 +157,7 @@ for file in files:
         locs.append(file['Key'])
 
 
-# In[21]:
+# In[13]:
 
 
 updated_locs = []
@@ -111,20 +171,34 @@ for loc in locs:
         continue
 
 
-# In[35]:
+# In[ ]:
 
 
 for loc,uloc in zip(locs,updated_locs):
     year = loc.split('/')[1]
+    if int(year)!=2019:
+        continue
     out_file = loc.split('/')[2] #has zip tag
     s3.download_file('capstone-storage',loc,Filename='unzipped/'+out_file)
     os.system('unzip unzipped/{} -d unzipped'.format(out_file))
-    os.system('python3 parse_patents.py -i unzipped/{}.xml'.format(uloc[:-4]))
+    os.system('python3 parse_patents.py -i Unzipped/{}.xml'.format(uloc[:-4]))
     os.remove('unzipped/{}.xml'.format(uloc[:-4]))
     os.remove('unzipped/{}'.format(out_file))
     df = dataframe_generator(uloc[:-4])
     shutil.rmtree('Unzipped/{}'.format(uloc[:-4]))
     df.to_csv('unzipped/{}.csv'.format(uloc[:-4]))
-    my_bucket.upload_file('unzipped/{}.csv'.format(uloc[:-4]), 'cleaned_data/{}/{}.csv'.format(year,uloc[:-4]))
+    my_bucket.upload_file('unzipped/{}.csv'.format(uloc[:-4]), 'cleaned_data_with_description/{}/{}.csv'.format(year,uloc[:-4]))
     os.remove('unzipped/{}.csv'.format(uloc[:-4]))
+
+
+# In[ ]:
+
+
+
+
+
+# In[ ]:
+
+
+
